dossier statistic-convert and cessStep duration

+++ modules/backend-dossiermgt/backend-dossiermgt-api/src/main/java/org/opencps/dossiermgt/service/ProcessStepLocalService.java
@@ -334,7 +334,7 @@ public interface ProcessStepLocalService extends BaseLocalService,
 	public ProcessStep updateProcessStep(long groupId, long processStepId,
 		String stepCode, String stepName, long serviceProcessId,
 		String sequenceNo, String dossierStatus, String dossierSubStatus,
-		int durationCount, String customProcessUrl, String stepInstruction,
+		double durationCount, String customProcessUrl, String stepInstruction,
 		String briefNote, boolean editable, String lockState,
 		Integer checkInput, ServiceContext context) throws PortalException;
